+++ blender_benchmark/manifest.yaml
